Fix MSVC warnings
This fixes all remaining MSVC type conversion warnings (both x64 and x32
build) and one "code unreachable" warning (by marking
HDF5ErrMapper::ToException as [[noreturn]]).

Also, in DataSpace::getMaxDimensions it fixes the H5S_UNLIMITED ->
DataSpace::UNLIMITED replacement to work on the correct vector.

Fixes #164

+++ include/highfive/bits/H5Converter_misc.hpp
@@ -46,7 +46,7 @@ inline bool is_1D(const std::vector<size_t>& dims)
 
 inline size_t compute_total_size(const std::vector<size_t>& dims)
 {
-    return std::accumulate(dims.begin(), dims.end(), 1ul,
+    return std::accumulate(dims.begin(), dims.end(), static_cast<size_t>(1),
                            std::multiplies<size_t>());
 }
